basic PSO algorithm functionality
You are now able to specify a config to construct a PSO system using
a factory to run an optimization. The algorithm is configurable
with the PSOConfig to allow for both minimization and maximization of
objective functions, initial particle solutions and velocities, inertia
and acceleration constants, as well as particle topologies. Currently
the termination criteria is limited to a number of iterations but that
will be expanded in a future commit.